third_party/blink/web_tests/external/wpt/html/semantics/the-button-element/command-and-commandfor/on-popover-invalid-behavior-expected.txt

This is a testharness.js-based test.
[FAIL] invoking (as show-modal) on popover does nothing
  assert_false: command event fired expected false got true
[FAIL] invoking (as show-modal) on open popover does nothing
  assert_false: command event fired expected false got true
[FAIL] invoking (as show-picker) on popover does nothing
  assert_false: command event fired expected false got true
[FAIL] invoking (as show-picker) on open popover does nothing
  assert_false: command event fired expected false got true
[FAIL] invoking (as open) on popover does nothing
  assert_false: command event fired expected false got true
[FAIL] invoking (as open) on open popover does nothing
  assert_false: command event fired expected false got true
[FAIL] invoking (as close) on popover does nothing
  assert_false: command event fired expected false got true
[FAIL] invoking (as close) on open popover does nothing
  assert_false: command event fired expected false got true
Harness: the test ran to completion.
